Create hash_chain.py
In this task your goal is to implement a hash table with lists chaining. You are already given the
number of buckets m and the hash function. It is a polynomial hash function
⎛
⎞
|S|−1
∑︁
h(S) = ⎝
S[i]x i mod p ⎠ mod m ,
i=0
where S[i] is the ASCII code of the i-th symbol of S, p = 1 000 000 007 and x = 263. Your program
should support the following kinds of queries:
∙ add string — insert string into the table. If there is already such string in the hash table, then
just ignore the query.
∙ del string — remove string from the table. If there is no such string in the hash table, then
just ignore the query.
∙ find string — output “yes" or “no" (without quotes) depending on whether the table contains
string or not.
∙ check i — output the content of the i-th list in the table. Use spaces to separate the elements of
the list. If i-th list is empty, output a blank line.