Largest even number possible by using one swap operation in given number
Given an odd number in the form of a string, the task is to make largest even number from the given number and you are allowed to do only one swap operation.

Examples :

Input : 1235785
Output :1535782
Swap 2 and 5.

Input :  536425
Output :  536524
Swap 4 and 5 to make the largest even number.

Find the first even number less than or equal to the odd number at last index.
If found, swap both values. Else swap with the last even value in the string.
If not possible to make even, print the given string.