Given a linked list of 0s, 1s and 2s, sort it.
Following steps can be used to sort the given linked list.

Traverse the list and count the number of 0s, 1s and 2s. Let the counts be n1, n2 and n3 respectively.
Traverse the list again, fill the first n1 nodes with 0, then n2 nodes with 1 and finally n3 nodes with 2.